test: 게시글 수정 기능 추가

+++ server/community/views.py
@@ -30,14 +30,23 @@ def review_create(request, movie_pk):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
-@api_view(['DELETE'])
+@api_view(['PUT', 'DELETE'])
 @authentication_classes([JSONWebTokenAuthentication]) # JWT가 유효한지 여부를 판단
 @permission_classes([IsAuthenticated]) # 인증 여부를 확인
-def review_delete(request, movie_pk, review_pk):
+def review_delete_update(request, movie_pk, review_pk):
     review = get_object_or_404(Review, pk=review_pk)
-
     if not request.user.user_reviews.filter(pk=review_pk).exists():
         return Response({'detail': '수정/삭제 권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
-    
-    review.delete()
-    return Response({ 'id': review_pk }, status=status.HTTP_204_NO_CONTENT)
+    try:
+        movie = Movie.objects.prefetch_related('reviews').filter(pk=movie_pk)[0]
+    except:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+
+    if request.method == 'PUT':
+        serializer = ReviewSerializer(review, data=request.data)
+        if serializer.is_valid(raise_exception=True):
+            serializer.save(movie=movie, user=request.user)
+            return Response(serializer.data)
+    else:
+        review.delete()
+        return Response({ 'id': review_pk }, status=status.HTTP_204_NO_CONTENT)
